requirements.txt
beautifulsoup4==4.11.1
lxml==4.9.1
requests
python-dotenv
tweety-ns==2.0.2
